fix: 修正CPU3参数字节序与通信声明（CPU2 V0.19.1.0 / CPU3 V0.20.1.0）
版本：
- CPU2：V0.19.1.0（历史提交编号）
- CPU3：V0.20.1.0（历史提交编号）

协议版本/兼容性：
- 涉及外部 Modbus、DSM 或瓦锡兰通信；接口口径以该历史提交的代码树为准。

本次修改：
- 修正 CPU3 参数写入时的寄存器字节序。
- 统一串口接收长度、协议处理声明和参数转发函数。

验证：
- 提交差异复核：通过；本次仅重写说明，代码树保持不变。

+++ LTD_DISPLAY_CPU3/Core/Src/usart.c
@@ -72,18 +72,18 @@ PUTCHAR_PROTOTYPE {
 #endif
 
 volatile uint8_t com2_rx_ready = 0;
-volatile uint8_t UART2_RX_LEN = 0;              // 接收一帧数据的长度
+volatile uint16_t UART2_RX_LEN = 0;              // 接收一帧数据的长度
 uint8_t UART2_RX_BUF[UART2_RX_BUF_SIZE] = { 0 };   // 接收数据缓冲区
 
 volatile uint8_t com3_rx_ready = 0;
-volatile uint8_t UART3_RX_LEN = 0;              // 接收一帧数据的长度
+volatile uint16_t UART3_RX_LEN = 0;              // 接收一帧数据的长度
 uint8_t UART3_RX_BUF[UART3_RX_BUF_SIZE] = { 0 };   // 接收数据缓冲区
 
-volatile uint8_t UART5_RX_LEN = 0;              // 接收一帧数据的长度
+volatile uint16_t UART5_RX_LEN = 0;              // 接收一帧数据的长度
 uint8_t UART5_RX_BUF[UART5_RX_BUF_SIZE] = { 0 };   // 接收数据缓冲区
 
 volatile uint8_t com1_rx_ready = 0;
-volatile uint8_t UART6_RX_LEN = 0;              // 接收一帧数据的长度
+volatile uint16_t UART6_RX_LEN = 0;              // 接收一帧数据的长度
 uint8_t UART6_RX_BUF[UART6_RX_BUF_SIZE] = { 0 };   // 接收数据缓冲区
 /* USER CODE END 0 */
